Add settings menu and move ContextMenuStrip logic into new class.
