+++ TorrentFlow/Program.cs
@@ -20,27 +20,11 @@ namespace TorrentFlow
             //Application.Run(new MainForm());
 
             using (var pi = new ProcessIcon("TorrentFlow"))
-            {
-                pi.SetContextMenuStrip(GetSettingsMenu());
-                
+            {           
                 pi.Display();
 
                 Application.Run();
             }
-        }
-        private static ContextMenuStrip GetSettingsMenu()
-        {
-            ContextMenuStrip menu = new ContextMenuStrip();
-            ToolStripMenuItem exitItem = new ToolStripMenuItem();
-            exitItem.Text = "Exit";
-            exitItem.Click += new EventHandler(ExitClick);
-            menu.Items.Add(exitItem);
-            return menu;
-        }
-
-        private static void ExitClick(object send, EventArgs e)
-        {
-            Application.Exit();
         }
     }
 }
